HOME-70 - Move alert switch to the agents list as it's cross-agent
