+++ client/modules/AgentsList/ControlPanel/ControlPanel.jsx
