+++ client/modules/AgentsList/ControlPanel/_styles.scss
