+++ client/modules/AgentsList/ControlPanel/index.jsx
