+++ client/modules/AgentsList/_styles.scss
@@ -1,3 +1,5 @@
+@import "./ControlPanel/styles";
+
 $color-pippin: #ffe2e0;
 $color-watusi: #ffd4cc;
 $color-persimmon: #ff7157;
