+++ client/modules/AgentsList/actionTypes.js
@@ -1,3 +1,10 @@
 export const DATA_FETCH = 'agents-list/fetch';
 export const DATA_FETCH_SUCCESS = 'agents-list/fetch-success';
 export const DATA_FETCH_ERROR = 'agents-list/fetch-fail';
+
+export const TOGGLE_ALERTS = 'dashboard/toggle-alerts';
+export const SET_ALERTS = 'dashboard/set-alerts';
+export const FETCH_ALERTS = 'dashboard/fetch-alerts';
+
+export const SEND_ALERT = 'dashboard/send-alert';
+
