+++ client/modules/AgentsList/actions.js
@@ -9,3 +9,21 @@ export const fetchDataFail = error => ({
   type: actionTypes.DATA_FETCH_ERROR,
   error,
 });
+
+export const toggleAlerts = () => ({
+  type: actionTypes.TOGGLE_ALERTS,
+});
+
+export const fetchAlerts = () => ({
+  type: actionTypes.FETCH_ALERTS,
+});
+
+export const sendAlert = () => ({
+  type: actionTypes.SEND_ALERT,
+});
+
+export const setAlerts = isAlerts => ({
+  type: actionTypes.SET_ALERTS,
+  isAlerts,
+});
+
