+++ client/modules/AgentsList/sagas.js
@@ -1,7 +1,8 @@
 import _ from 'lodash';
 import { delay } from 'redux-saga';
-import { put, call, fork } from 'redux-saga/effects';
+import {  put, call, fork, takeEvery } from 'redux-saga/effects';
 import * as actions from './actions';
+import * as actionTypes from './actionTypes';
 
 function getData() {
   return fetch('/api/agents')
@@ -19,6 +20,12 @@ function getData() {
     .catch(e => e);
 }
 
+function callSendAlert() {
+  return fetch('/api/sendalert', { method: 'POST' })
+    .then(response => response.json())
+    .catch(() => 'Send alert failed');
+}
+
 function* fetchData() {
   while (true) {
     const agents = yield call(getData);
@@ -33,9 +40,51 @@ function* fetchData() {
   }
 }
 
+function* sendAlert() {
+  yield call(callSendAlert);
+}
+
+function callToggleAlerts() {
+  return fetch('/api/alerts', { method: 'POST' })
+    .then(response => response.json())
+    .catch(() => 'Toggling alerts failed');
+}
+
+function* toggleAlerts() {
+  const data = yield call(callToggleAlerts);
+
+  if (_.isObject(data)) {
+    const isAlerts = data.isAlerts === 'true';
+
+    yield put(actions.setAlerts(isAlerts));
+  }
+}
+
+function callAlerts() {
+  return fetch('/api/alerts')
+    .then(response => response.json())
+    .catch(() => 'Toggling alerts failed');
+}
+
+function* fetchAlerts() {
+  const data = yield call(callAlerts);
+
+  if (_.isObject(data)) {
+    const isAlerts = data.isAlerts === 'true';
+
+    yield put(actions.setAlerts(isAlerts));
+  } else {
+    // @TODO: Figure out how to handle failed actions on the UI side
+  }
+}
+
 function* root() {
   yield [
     fork(fetchData),
+    fork(fetchAlerts),
+    takeEvery(actionTypes.TOGGLE_ALERTS, toggleAlerts),
+    takeEvery(actionTypes.FETCH_ALERTS, fetchAlerts),
+    takeEvery(actionTypes.SEND_ALERT, sendAlert),
   ];
 }
