+++ client/modules/Dashboard/Dashboard.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ControlPanel from './ControlPanel';
 import TemperaturePanel from './TemperaturePanel';
 import SoundPanel from './SoundPanel';
 import CurrentPanel from './CurrentPanel';
@@ -16,9 +15,6 @@ const Dashboard = (props) => {
         </div>
       )}
 
-      <div className="dashboard__cell dashboard__cell--full">
-        <ControlPanel />
-      </div>
       <div className="dashboard__cell dashboard__cell--full">
         <CurrentPanel />
       </div>
