+++ client/modules/Dashboard/_styles.scss
@@ -1,4 +1,3 @@
-@import "ControlPanel/styles";
 @import "TemperaturePanel/styles";
 @import "SoundPanel/styles";
 @import "CurrentPanel/styles";
