+++ client/modules/Dashboard/actionTypes.js
@@ -1,9 +1,3 @@
 export const DATA_FETCH = 'dashboard/fetch';
 export const DATA_FETCH_SUCCESS = 'dashboard/fetch-success';
 export const DATA_FETCH_ERROR = 'dashboard/fetch-fail';
-
-export const TOGGLE_ALERTS = 'dashboard/toggle-alerts';
-export const SET_ALERTS = 'dashboard/set-alerts';
-export const FETCH_ALERTS = 'dashboard/fetch-alerts';
-
-export const SEND_ALERT = 'dashboard/send-alert';
