+++ client/modules/Dashboard/reducers.js
@@ -30,9 +30,6 @@ export default function reducer(state = defaultState, action) {
     case actionTypes.DATA_FETCH_ERROR:
       return Object.assign({}, state, { error: action.error });
 
-    case actionTypes.SET_ALERTS:
-      return Object.assign({}, state, { isAlerts: action.isAlerts });
-
     default:
       return state;
   }
