+++ client/modules/Dashboard/sagas.js
@@ -2,51 +2,6 @@ import _ from 'lodash';
 import { delay } from 'redux-saga';
 import { put, call, fork, takeEvery } from 'redux-saga/effects';
 import * as actions from './actions';
-import * as actionTypes from './actionTypes';
-
-function callSendAlert() {
-  return fetch('/api/sendalert', { method: 'POST' })
-    .then(response => response.json())
-    .catch(() => 'Send alert failed');
-}
-
-function* sendAlert() {
-  yield call(callSendAlert);
-}
-
-function callToggleAlerts() {
-  return fetch('/api/alerts', { method: 'POST' })
-    .then(response => response.json())
-    .catch(() => 'Toggling alerts failed');
-}
-
-function* toggleAlerts() {
-  const data = yield call(callToggleAlerts);
-
-  if (_.isObject(data)) {
-    const isAlerts = data.isAlerts === 'true';
-
-    yield put(actions.setAlerts(isAlerts));
-  }
-}
-
-function callAlerts() {
-  return fetch('/api/alerts')
-    .then(response => response.json())
-    .catch(() => 'Toggling alerts failed');
-}
-
-function* fetchAlerts() {
-  const data = yield call(callAlerts);
-
-  if (_.isObject(data)) {
-    const isAlerts = data.isAlerts === 'true';
-
-    yield put(actions.setAlerts(isAlerts));
-  } else {
-    // @TODO: Figure out how to handle failed actions on the UI side
-  }
-}
 
 function getData(agentId) {
   return fetch(`/api/agents/${agentId}`)
@@ -91,10 +46,6 @@ function* fetchData() {
 function* root() {
   yield [
     fork(fetchData),
-    fork(fetchAlerts),
-    takeEvery(actionTypes.TOGGLE_ALERTS, toggleAlerts),
-    takeEvery(actionTypes.FETCH_ALERTS, fetchAlerts),
-    takeEvery(actionTypes.SEND_ALERT, sendAlert),
   ];
 }
