+++ views/dashboard.html
@@ -1,9 +1,6 @@
 {{define "body"}}
 <div class="c-page">
     <section>
-        <header>
-            Available agents
-        </header>
         <article class="js-agents-list">
         </article>
     </section>
