+++ views/view.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
     <title>
-        {{.Title}}
+        Smarthome - {{.Title}}
     </title>
     <link rel="stylesheet" href="/static/styles.css">
 </head>
